pyproject.toml
[tool.poetry]
name = "psc"
version = "0.0.0"
description = "PyScript Collective"
authors = ["Paul Everitt <pauleveritt@me.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/pauleveritt/psc"
repository = "https://github.com/pauleveritt/psc"
documentation = "https://psc.readthedocs.io"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.urls]
Changelog = "https://github.com/pauleveritt/psc/releases"

[tool.poetry.dependencies]
python = "^3.7"
click = ">=8.0.1"
nox = "^2022.1.7"
typer = { extras = ["all"], version = "^0.6.1" }
starlette = "^0.20.4"
uvicorn = "^0.18.2"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = { extras = ["toml"], version = ">=6.2" }
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = { extras = ["colors"], version = ">=0.15.10" }
myst-parser = { version = ">=0.16.1" }
requests = "^2.28.1"
pytest-playwright = "^0.3.0"
pytest-asyncio-cooperative = "^0.28.0"
nox-poetry = "^1.0.1"
beautifulsoup4 = "^4.11.1"
html5lib = "^1.1"
types-beautifulsoup4 = "^4.11.4"
Jinja2 = "^3.1.2"
types-requests = "^2.28.10"

[tool.poetry.scripts]
psc = "psc.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["psc"]

[tool.coverage.report]
show_missing = true
fail_under = 95

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
files = ["src/psc", "tests"]
exclude = [".nox", "docs","static","pyodide"]
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    'uvicorn',
    'jinja2',
    'frontmatter'
]
ignore_missing_imports = true


[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "unit",
    "shallow",
    "full: marks Playwright as slow (deselect with '-m \"not full\"')",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
